# src/swagger_files/QuickCode.Demo.UserManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Demo - UserManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/user-manager-module/kafka-events/{kafkaEventsTopicName}/topic-workflows/{topicWorkflowsId}": {
      "get": {
        "tags": [
          "KafkaEvents"
        ],
        "operationId": "api-user-manager-module-kafka-events--kafkaEventsTopicName--topic-workflows--topicWorkflowsId--get",
        "parameters": [
          {
            "name": "kafkaEventsTopicName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topicWorkflowsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KafkaEventsTopicWorkflows_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/topic-workflows/count": {
      "get": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows-count-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/topic-workflows/{id}": {
      "get": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicWorkflowsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicWorkflowsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicWorkflowsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TopicWorkflowsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/topic-workflows/get-workflows/{topicWorkflowsKafkaEventsTopicName}": {
      "get": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows-get-workflows--topicWorkflowsKafkaEventsTopicName--get",
        "parameters": [
          {
            "name": "topicWorkflowsKafkaEventsTopicName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopicWorkflowsGetWorkflowsResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/topic-workflows/get-topic-workflows/{kafkaEventsTopicName}/{apiMethodDefinitionsHttpMethod}": {
      "get": {
        "tags": [
          "TopicWorkflows"
        ],
        "operationId": "api-user-manager-module-topic-workflows-get-topic-workflows--kafkaEventsTopicName---apiMethodDefinitionsHttpMethod--get",
        "parameters": [
          {
            "name": "kafkaEventsTopicName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiMethodDefinitionsHttpMethod",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopicWorkflowsGetTopicWorkflowsResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "KafkaEventsTopicWorkflows_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kafkaEventsTopicName": {
            "type": "string",
            "nullable": true
          },
          "workflowContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TopicWorkflowsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kafkaEventsTopicName": {
            "type": "string",
            "nullable": true
          },
          "workflowContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TopicWorkflowsGetTopicWorkflowsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kafkaEventsTopicName": {
            "type": "string",
            "nullable": true
          },
          "workflowContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TopicWorkflowsGetWorkflowsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kafkaEventsTopicName": {
            "type": "string",
            "nullable": true
          },
          "workflowContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}